test(e2e): split monolithic e2e specs into focused files by feature area - for-root: initialization, auth-api-basic, auth-api-register, auth-api-login,   auth-api-reset-password, auth-api-broadcast, websockets-auth-basic,   websockets-auth-register, websockets-auth-login, websockets-auth-reset-password,   websockets-auth-broadcast - for-feature: base-entity, soft-deletable-entity, websockets, http-broadcast - removed: dynamic-api-for-feature.e2e-spec.ts, dynamic-api-for-root.e2e-spec.ts
