--- libs/dynamic-api/test/dynamic-api-for-feature.e2e-spec.ts
@@ -1,1363 +1,0 @@
-import { INestApplication } from '@nestjs/common';
-import { Prop, Schema } from '@nestjs/mongoose';
-import { Test } from '@nestjs/testing';
-import mongoose, { Connection } from 'mongoose';
-import {
-  BaseEntity,
-  DynamicApiForFeatureOptions,
-  DynamicApiForRootOptions,
-  DynamicApiModule,
-  DynamicAPISchemaOptions,
-  SoftDeletableEntity,
-} from '../src';
-import { closeTestingApp, createTestingApp, handleSocketBroadcast, server, TestGateway, TestSocketAdapter } from './e2e.setup';
-import 'dotenv/config';
-import { getModelFromEntity } from './utils';
-
-describe('DynamicApiModule forFeature (e2e)', () => {
-  const uri = process.env.MONGO_DB_URL;
-
-  const initApp = async <Entity extends BaseEntity, UserEntity extends BaseEntity>(
-    forFeatureOptions: DynamicApiForFeatureOptions<Entity>,
-    forRootOptions: DynamicApiForRootOptions<UserEntity> = {},
-    initFixtures?: (connection: Connection) => Promise<void>,
-    initMainCb?: (app: any) => Promise<void>,
-    testGateway?: boolean,
-  ) => {
-    const moduleRef = await Test.createTestingModule({
-      imports: [
-        DynamicApiModule.forRoot(uri, forRootOptions),
-        DynamicApiModule.forFeature(forFeatureOptions),
-      ],
-      providers: testGateway ? [TestGateway] : [],
-    }).compile();
-
-    await createTestingApp(moduleRef, initFixtures, initMainCb);
-  };
-
-  beforeEach(() => {
-    DynamicApiModule.state['resetState']();
-  });
-
-  afterEach(async () => {
-    await closeTestingApp(mongoose.connections);
-  });
-
-  describe('Entity extends BaseEntity', () => {
-    @DynamicAPISchemaOptions({
-      indexes: [{ fields: { name: 1 }, options: { unique: true } }],
-    })
-    @Schema({ collection: 'test-entities' })
-    class TestEntity extends BaseEntity {
-      @Prop({ type: String, required: true })
-      name: string;
-
-      @Prop({ type: String })
-      group: string;
-    }
-
-    beforeEach(async () => {
-      const fixtures = async (_: Connection) => {
-        const model = await getModelFromEntity(TestEntity);
-        await model.insertMany([
-          { name: 'test1' },
-          { name: 'test2' },
-        ]);
-      };
-
-      await initApp(
-        {
-          entity: TestEntity,
-          controllerOptions: {
-            path: 'test-entities',
-          },
-        },
-        {},
-        fixtures,
-      );
-    });
-
-    describe('GET /test-entities', () => {
-      it('should return all test entities', async () => {
-        const { body, status } = await server.get('/test-entities');
-
-        expect(status).toBe(200);
-        expect(body).toEqual([
-          {
-            id: expect.any(String),
-            name: 'test1',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-          {
-            id: expect.any(String),
-            name: 'test2',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-    });
-
-    describe('GET /test-entities/:id', () => {
-      let entities: TestEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should return a test entity by id', async () => {
-        const { body, status } = await server.get(`/test-entities/${entities[0].id}`);
-
-        expect(status).toBe(200);
-        expect(body).toEqual(entities[0]);
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.get(`/test-entities/123`);
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestEntity not found');
-      });
-    });
-
-    describe('POST /test-entities/many', () => {
-      it('should create many test entities', async () => {
-        const { body, status } = await server.post('/test-entities/many', {
-          list: [
-            { name: 'test3' },
-            { name: 'test4' },
-          ],
-        });
-
-        expect(status).toBe(201);
-        expect(body).toEqual([
-          {
-            id: expect.any(String),
-            name: 'test3',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-          {
-            id: expect.any(String),
-            name: 'test4',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-
-      // it('should return 409 if entity has duplicated key', async () => {
-      //   const { body, status } = await server.post('/test-entities/many', {
-      //     list: [
-      //       { name: 'test1' },
-      //       { name: 'test2' },
-      //     ],
-      //   });
-      //
-      //   expect(status).toBe(409);
-      //   expect(body).toHaveProperty('error', 'Conflict');
-      //   expect(body).toHaveProperty('statusCode', 409);
-      //   expect(body).toHaveProperty('message', 'name \'test1\' is already used');
-      // });
-    });
-
-    describe('POST /test-entities', () => {
-      it('should create a test entity', async () => {
-        const { body, status } = await server.post('/test-entities', { name: 'test3' });
-
-        expect(status).toBe(201);
-        expect(body).toEqual({
-          id: expect.any(String),
-          name: 'test3',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      // it('should return 409 if entity has duplicated key', async () => {
-      //   const { body, status } = await server.post('/test-entities', { name: 'test1' });
-      //
-      //   expect(status).toBe(409);
-      //   expect(body).toHaveProperty('error', 'Conflict');
-      //   expect(body).toHaveProperty('statusCode', 409);
-      //   expect(body).toHaveProperty('message', 'name \'test1\' is already used');
-      // });
-    });
-
-    describe('PUT /test-entities/:id', () => {
-      let entityToReplace: TestEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToReplace = body.pop();
-      });
-
-      it('should replace a test entity by id', async () => {
-        const { body, status } = await server.put(`/test-entities/${entityToReplace.id}`, { name: 'test-replaced' });
-
-        expect(status).toBe(200);
-        expect(body).toEqual({
-          id: expect.any(String),
-          name: 'test-replaced',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.put(`/test-entities/123`, { name: 'test1-updated' });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestEntity not found');
-      });
-    });
-
-    describe('PATCH /test-entities', () => {
-      let entities: TestEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should update many test entities', async () => {
-        const { body, status } = await server.patch('/test-entities', {
-          group: 'many-patched',
-        }, { query: { ids: entities.map(({ id }) => id) } });
-
-        expect(status).toBe(200);
-        expect(body).toEqual([
-          {
-            ...entities[0],
-            group: 'many-patched',
-            updatedAt: expect.any(String),
-          },
-          {
-            ...entities[1],
-            group: 'many-patched',
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.patch('/test-entities', {
-          group: 'many-patched',
-        }, { query: { ids: ['123'] } });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestEntity not found');
-      });
-    });
-
-    describe('PATCH /test-entities/:id', () => {
-      let entityToUpdate: TestEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToUpdate = body.pop();
-      });
-
-      it('should update a test entity by id', async () => {
-        const { body, status } = await server.patch(`/test-entities/${entityToUpdate.id}`, { name: 'test-updated' });
-
-        expect(status).toBe(200);
-        expect(body).toEqual({
-          id: expect.any(String),
-          name: 'test-updated',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.patch(`/test-entities/123`, { name: 'test1-updated' });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestEntity not found');
-      });
-    });
-
-    describe('DELETE /test-entities', () => {
-      let entities: TestEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should delete many test entities', async () => {
-        const { body, status } = await server.delete(
-          '/test-entities',
-          { query: { ids: entities.map(({ id }) => id) } },
-        );
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 2 });
-
-        const model = await getModelFromEntity(TestEntity);
-        const updatedList = await model.find().lean().exec();
-
-        expect(updatedList).toEqual([]);
-      });
-
-      it('should return 200 with 0 for deletedCount if one of entities is not found', async () => {
-        const { body, status } = await server.delete('/test-entities', { query: { ids: [entities[0].id, '123'] } });
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 0 });
-      });
-    });
-
-    describe('DELETE /test-entities/:id', () => {
-      let entityToDelete: TestEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToDelete = body.pop();
-      });
-
-      it('should delete a test entity by id', async () => {
-        const { body, status } = await server.delete(`/test-entities/${entityToDelete.id}`);
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 1 });
-      });
-
-      it('should return 200 with deletedCount at 0 if entity not found', async () => {
-        const { body, status } = await server.delete(`/test-entities/123`);
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 0 });
-      });
-    });
-
-    describe('POST /test-entities/duplicate', () => {
-      let entities: TestEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should duplicate test entities', async () => {
-        const { body, status } = await server.post('/test-entities/duplicate', {
-          name: 'test1-duplicated',
-          group: 'duplicated',
-        }, { query: { ids: [entities[0].id] } });
-
-        expect(status).toBe(201);
-        expect(body).toEqual([
-          {
-            ...entities[0],
-            id: expect.any(String),
-            name: 'test1-duplicated',
-            group: 'duplicated',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-
-      it('should return 404 if one of entities is not found', async () => {
-        const { body, status } = await server.post(`/test-entities/duplicate`, {
-          group: 'duplicated',
-        }, { query: { ids: [entities[0].id, '123'] } });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestEntity not found');
-      });
-    });
-
-    describe('POST /test-entities/duplicate/:id', () => {
-      let entityToDuplicate: TestEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToDuplicate = body.pop();
-      });
-
-      it('should duplicate a test entity by id', async () => {
-        const { body, status } = await server.post(`/test-entities/duplicate/${entityToDuplicate.id}`, {
-          name: 'test2-duplicated',
-        });
-
-        expect(status).toBe(201);
-        expect(body).toEqual({
-          ...entityToDuplicate,
-          id: expect.any(String),
-          name: 'test2-duplicated',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.post(`/test-entities/duplicate/123`, {
-          name: 'test1-duplicated',
-          group: 'duplicated',
-        });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestEntity not found');
-      });
-    });
-  });
-
-  describe('Entity extends SoftDeletableEntity', () => {
-    @DynamicAPISchemaOptions({
-      indexes: [{ fields: { name: 1 }, options: { unique: true, partialFilterExpression: { deletedAt: { $eq: null } } } }],
-    })
-    @Schema({ collection: 'test-entities' })
-    class TestDeletableEntity extends SoftDeletableEntity {
-      @Prop({ type: String, required: true })
-      name: string;
-
-      @Prop({ type: String })
-      group: string;
-    }
-
-    beforeEach(async () => {
-      const fixtures = async (_: Connection) => {
-        const model = await getModelFromEntity(TestDeletableEntity);
-        await model.insertMany([
-          { name: 'test1' },
-          { name: 'test2' },
-        ]);
-      };
-
-      await initApp(
-        {
-          entity: TestDeletableEntity,
-          controllerOptions: {
-            path: 'test-entities',
-          },
-        },
-        {},
-        fixtures,
-      );
-    });
-
-    describe('GET /test-entities', () => {
-      it('should return all test entities', async () => {
-        const { body, status } = await server.get('/test-entities');
-
-        expect(status).toBe(200);
-        expect(body).toEqual([
-          {
-            id: expect.any(String),
-            name: 'test1',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-          {
-            id: expect.any(String),
-            name: 'test2',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-    });
-
-    describe('GET /test-entities/:id', () => {
-      let entities: TestDeletableEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should return a test entity by id', async () => {
-        const { body, status } = await server.get(`/test-entities/${entities[0].id}`);
-
-        expect(status).toBe(200);
-        expect(body).toEqual(entities[0]);
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.get(`/test-entities/123`);
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestDeletableEntity not found');
-      });
-    });
-
-    describe('POST /test-entities/many', () => {
-      it('should create many test entities', async () => {
-        const { body, status } = await server.post('/test-entities/many', {
-          list: [
-            { name: 'test3' },
-            { name: 'test4' },
-          ],
-        });
-
-        expect(status).toBe(201);
-        expect(body).toEqual([
-          {
-            id: expect.any(String),
-            name: 'test3',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-          {
-            id: expect.any(String),
-            name: 'test4',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-    });
-
-    describe('POST /test-entities', () => {
-      it('should create a test entity', async () => {
-        const { body, status } = await server.post('/test-entities', { name: 'test3' });
-
-        expect(status).toBe(201);
-        expect(body).toEqual({
-          id: expect.any(String),
-          name: 'test3',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      // it('should return 409 if entity has duplicated key', async () => {
-      //   const { body, status } = await server.post('/test-entities', { name: 'test1' });
-      //
-      //   expect(status).toBe(409);
-      //   expect(body).toHaveProperty('error', 'Conflict');
-      //   expect(body).toHaveProperty('statusCode', 409);
-      //   expect(body).toHaveProperty('message', 'name \'test1\' is already used');
-      // });
-    });
-
-    describe('PUT /test-entities/:id', () => {
-      let entityToReplace: TestDeletableEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToReplace = body.pop();
-      });
-
-      it('should replace a test entity by id', async () => {
-        const { body, status } = await server.put(`/test-entities/${entityToReplace.id}`, { name: 'test-replaced' });
-
-        expect(status).toBe(200);
-        expect(body).toEqual({
-          id: expect.any(String),
-          name: 'test-replaced',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.put(`/test-entities/123`, { name: 'test1-updated' });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestDeletableEntity not found');
-      });
-    });
-
-    describe('PATCH /test-entities', () => {
-      let entities: TestDeletableEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should update many test entities', async () => {
-        const { body, status } = await server.patch('/test-entities', {
-          group: 'many-patched',
-        }, { query: { ids: entities.map(({ id }) => id) } });
-
-        expect(status).toBe(200);
-        expect(body).toEqual([
-          {
-            ...entities[0],
-            group: 'many-patched',
-            updatedAt: expect.any(String),
-          },
-          {
-            ...entities[1],
-            group: 'many-patched',
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.patch('/test-entities', {
-          group: 'many-patched',
-        }, { query: { ids: ['123'] } });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestDeletableEntity not found');
-      });
-    });
-
-    describe('PATCH /test-entities/:id', () => {
-      let entityToUpdate: TestDeletableEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToUpdate = body.pop();
-      });
-
-      it('should update a test entity by id', async () => {
-        const { body, status } = await server.patch(`/test-entities/${entityToUpdate.id}`, { name: 'test-updated' });
-
-        expect(status).toBe(200);
-        expect(body).toEqual({
-          id: expect.any(String),
-          name: 'test-updated',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.patch(`/test-entities/123`, { name: 'test1-updated' });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestDeletableEntity not found');
-      });
-    });
-
-    describe('DELETE /test-entities', () => {
-      let entities: TestDeletableEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should delete many test entities', async () => {
-        const { body, status } = await server.delete(
-          '/test-entities',
-          { query: { ids: entities.map(({ id }) => id) } },
-        );
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 2 });
-
-        const model = await getModelFromEntity(TestDeletableEntity);
-        const updatedList = await model.find().lean().exec();
-
-        expect(updatedList).toEqual([
-          {
-            _id: expect.any(Object),
-            __v: expect.any(Number),
-            name: 'test1',
-            isDeleted: true,
-            createdAt: expect.any(Date),
-            updatedAt: expect.any(Date),
-            deletedAt: expect.any(Date),
-          },
-          {
-            _id: expect.any(Object),
-            __v: expect.any(Number),
-            name: 'test2',
-            isDeleted: true,
-            createdAt: expect.any(Date),
-            updatedAt: expect.any(Date),
-            deletedAt: expect.any(Date),
-          },
-        ]);
-      });
-
-      it('should return 200 with 0 for deletedCount if one of entities is not found', async () => {
-        const { body, status } = await server.delete('/test-entities', { query: { ids: [entities[0].id, '123'] } });
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 0 });
-      });
-    });
-
-    describe('DELETE /test-entities/:id', () => {
-      let entityToDelete: TestDeletableEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToDelete = body.pop();
-      });
-
-      it('should delete a test entity by id', async () => {
-        const { body, status } = await server.delete(`/test-entities/${entityToDelete.id}`);
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 1 });
-      });
-
-      it('should return 200 with deletedCount at 0 if entity not found', async () => {
-        const { body, status } = await server.delete(`/test-entities/123`);
-
-        expect(status).toBe(200);
-        expect(body).toEqual({ deletedCount: 0 });
-      });
-    });
-
-    describe('POST /test-entities/duplicate', () => {
-      let entities: TestDeletableEntity[];
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entities = body;
-      });
-
-      it('should duplicate test entities', async () => {
-        const { body, status } = await server.post('/test-entities/duplicate', {
-          name: 'test1-duplicated',
-          group: 'duplicated',
-        }, { query: { ids: [entities[0].id] } });
-
-        expect(status).toBe(201);
-        expect(body).toEqual([
-          {
-            ...entities[0],
-            id: expect.any(String),
-            name: 'test1-duplicated',
-            group: 'duplicated',
-            createdAt: expect.any(String),
-            updatedAt: expect.any(String),
-          },
-        ]);
-      });
-
-      it('should return 404 if one of entities is not found', async () => {
-        const { body, status } = await server.post(`/test-entities/duplicate`, {
-          group: 'duplicated',
-        }, { query: { ids: [entities[0].id, '123'] } });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestDeletableEntity not found');
-      });
-    });
-
-    describe('POST /test-entities/duplicate/:id', () => {
-      let entityToDuplicate: TestDeletableEntity;
-
-      beforeEach(async () => {
-        const { body } = await server.get('/test-entities');
-        entityToDuplicate = body.pop();
-      });
-
-      it('should duplicate a test entity by id', async () => {
-        const { body, status } = await server.post(`/test-entities/duplicate/${entityToDuplicate.id}`, {
-          name: 'test2-duplicated',
-        });
-
-        expect(status).toBe(201);
-        expect(body).toEqual({
-          ...entityToDuplicate,
-          id: expect.any(String),
-          name: 'test2-duplicated',
-          createdAt: expect.any(String),
-          updatedAt: expect.any(String),
-        });
-      });
-
-      it('should return 404 if entity not found', async () => {
-        const { body, status } = await server.post(`/test-entities/duplicate/123`, {
-          name: 'test1-duplicated',
-          group: 'duplicated',
-        });
-
-        expect(status).toBe(404);
-        expect(body).toHaveProperty('error', 'Not Found');
-        expect(body).toHaveProperty('statusCode', 404);
-        expect(body).toHaveProperty('message', 'TestDeletableEntity not found');
-      });
-    });
-  });
-
-  describe('Websockets', () => {
-    @Schema({ collection: 'users' })
-    class UserEntity extends BaseEntity {
-      @Prop({ type: String, required: true })
-      email: string;
-
-      @Prop({ type: String, required: true })
-      password: string;
-
-      @Prop({ type: Boolean, default: false })
-      isAdmin?: boolean;
-    }
-
-    @Schema({ collection: 'products' })
-    class ProductEntity extends BaseEntity {
-      @Prop({ type: String, required: true })
-      name: string;
-
-      @Prop({ type: Boolean, default: false })
-      inStock: boolean;
-
-      @Prop({ type: String })
-      status?: string;
-    }
-
-    const customEvent = 'product-custom-event';
-    let adminAccessToken: string;
-    let userAccessToken: string;
-    let firstProduct: ProductEntity;
-    let secondProduct: ProductEntity;
-
-    beforeEach(async () => {
-      const fixtures = async (_: Connection) => {
-        const model = await getModelFromEntity(ProductEntity);
-        await model.insertMany([
-          { name: 'mayo', inStock: true },
-          { name: 'ketchup' },
-        ]).then(([mayoProduct, ketchupProduct]) => {
-          firstProduct = mayoProduct as unknown as ProductEntity;
-          secondProduct = ketchupProduct as unknown as ProductEntity;
-        });
-      };
-
-      await initApp(
-        {
-          entity: ProductEntity,
-          controllerOptions: {
-            apiTag: 'Product',
-            path: 'products',
-          },
-          routes: [
-            { type: 'GetMany', isPublic: true },
-            {
-              type: 'CreateMany',
-              webSocket: true,
-              broadcast: {
-                enabled: (_, user) => user.isAdmin === true,
-              },
-            },
-            {
-              type: 'CreateOne',
-              webSocket: true,
-              broadcast: { enabled: true },
-            },
-            {
-              type: 'UpdateMany',
-              webSocket: true,
-              broadcast: {
-                enabled: false,
-                eventName: customEvent,
-              },
-            },
-            {
-              type: 'UpdateOne',
-              webSocket: true,
-              broadcast: {
-                enabled: true,
-                eventName: customEvent,
-              },
-            },
-            {
-              type: 'ReplaceOne',
-              webSocket: true,
-              broadcast: {
-                enabled: true,
-              },
-            },
-            {
-              type: 'DuplicateOne',
-              webSocket: true,
-              broadcast: {
-                enabled: (_, user) => user.isAdmin === true,
-                eventName: customEvent,
-              },
-            },
-            {
-              type: 'DuplicateMany',
-              webSocket: true,
-              broadcast: {
-                enabled: true,
-              },
-            },
-            {
-              type: 'DeleteOne',
-              webSocket: true,
-              broadcast: {
-                enabled: true,
-              },
-            },
-            {
-              type: 'DeleteMany',
-              webSocket: true,
-              broadcast: {
-                enabled: true,
-              },
-            },
-          ],
-        },
-        {
-          useAuth: {
-            userEntity: UserEntity,
-            login: {
-              additionalFields: ['isAdmin'],
-            },
-            webSocket: true,
-          },
-        },
-        fixtures,
-        async (app: INestApplication) => {
-          app.useWebSocketAdapter(new TestSocketAdapter(app));
-        },
-        true,
-      );
-
-      adminAccessToken = (await server.emit(
-        'auth-register',
-        { email: 'unit@test.co', password: 'test', isAdmin: true },
-      )).accessToken;
-      userAccessToken = (await server.emit(
-        'auth-register',
-        { email: 'toto@test.co', password: 'test' },
-      )).accessToken;
-    });
-
-    it('[CreateOne] should broadcast created product to all clients', async () => {
-      await server.emit('create-one-product', { name: 'mustard', inStock: true }, { accessToken: adminAccessToken, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'create-one-product',
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            name: 'mustard',
-            inStock: true,
-          }),
-        ]),
-      });
-    });
-
-    it('[CreateMany] should broadcast created products to all clients when admin', async () => {
-      await server.emit('create-many-product', {
-        list: [
-          { name: 'mustard', inStock: true },
-          { name: 'relish', inStock: false },
-        ],
-      }, { accessToken: adminAccessToken, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'create-many-product',
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            name: 'mustard',
-            inStock: true,
-          }),
-          expect.objectContaining({
-            name: 'relish',
-            inStock: false,
-          }),
-        ]),
-      });
-    });
-
-    it('[CreateMany] should NOT broadcast created products when user is not admin', async () => {
-      await server.emit('create-many-product', {
-        list: [
-          { name: 'pickles', inStock: true },
-          { name: 'onions', inStock: false },
-        ],
-      }, { accessToken: userAccessToken });
-
-      expect(handleSocketBroadcast).not.toHaveBeenCalled();
-    });
-
-    it('[UpdateOne] should broadcast updated product to all clients with custom event name', async () => {
-      await server.emit('update-one-product', {
-        id: firstProduct.id,
-        name: 'updated mayo',
-        inStock: false,
-      }, { accessToken: adminAccessToken, broadcastEvent: customEvent, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: customEvent,
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            id: firstProduct.id,
-            name: 'updated mayo',
-            inStock: false,
-          }),
-        ]),
-      });
-    });
-
-    it('[UpdateMany] should NOT broadcast updated products when broadcast is disabled', async () => {
-      const { body: products } = await server.get('/products');
-
-      await server.emit('update-many-product', {
-        ids: products.map(p => p.id),
-        status: 'updated',
-      }, { accessToken: adminAccessToken });
-
-      expect(handleSocketBroadcast).not.toHaveBeenCalled();
-    });
-
-    it('[DeleteOne] should broadcast deleted product id to all clients', async () => {
-      await server.emit('delete-one-product', {
-        id: firstProduct.id,
-      }, { accessToken: adminAccessToken, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'delete-one-product',
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            id: firstProduct.id,
-          }),
-        ]),
-      });
-    });
-
-    it('[ReplaceOne] should broadcast replaced product to all clients', async () => {
-      await server.emit('replace-one-product', {
-        id: firstProduct.id,
-        name: 'replaced mayo',
-        inStock: false,
-      }, { accessToken: adminAccessToken, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'replace-one-product',
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            id: firstProduct.id,
-            name: 'replaced mayo',
-            inStock: false,
-          }),
-        ]),
-      });
-    });
-
-    it('[DuplicateOne] should broadcast duplicated product to all clients with custom event name when admin', async () => {
-      await server.emit('duplicate-one-product', {
-        id: firstProduct.id,
-        name: 'duplicated mayo',
-      }, { accessToken: adminAccessToken, broadcastEvent: customEvent, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: customEvent,
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            name: 'duplicated mayo',
-            inStock: firstProduct.inStock,
-          }),
-        ]),
-      });
-    });
-
-    it('[DuplicateOne] should NOT broadcast duplicated product when user is not admin', async () => {
-      await server.emit('duplicate-one-product', {
-        id: firstProduct.id,
-        name: 'duplicated mayo user',
-      }, { accessToken: userAccessToken });
-
-      expect(handleSocketBroadcast).not.toHaveBeenCalled();
-    });
-
-    it('[DuplicateMany] should broadcast duplicated products to all clients', async () => {
-      await server.emit('duplicate-many-product', {
-        ids: [secondProduct.id, firstProduct.id],
-        status: 'duplicated',
-      }, { accessToken: adminAccessToken, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'duplicate-many-product',
-        data: expect.arrayContaining([
-          expect.objectContaining({
-            id: expect.any(String),
-            name: secondProduct.name,
-            inStock: secondProduct.inStock,
-            status: 'duplicated',
-          }),
-          expect.objectContaining({
-            id: expect.any(String),
-            name: firstProduct.name,
-            inStock: firstProduct.inStock,
-            status: 'duplicated',
-          }),
-        ]),
-      });
-    });
-
-    it('[DeleteMany] should broadcast deleted product ids to all clients', async () => {
-      const products = [firstProduct, secondProduct];
-
-      await server.emit('delete-many-product', {
-        ids: products.map(p => p.id),
-      }, { accessToken: adminAccessToken, expectBroadcast: true });
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'delete-many-product',
-        data: expect.arrayContaining(
-          products.map(p => ({ id: p.id })),
-        ),
-      });
-    });
-  });
-
-  describe('HTTP Broadcast', () => {
-    @Schema({ collection: 'hb_products' })
-    class HbProductEntity extends BaseEntity {
-      @Prop({ type: String, required: true })
-      name: string;
-
-      @Prop({ type: Boolean, default: false })
-      inStock: boolean;
-
-      @Prop({ type: String })
-      status?: string;
-    }
-
-    const customEvent = 'hb-product-custom-event';
-    let firstProduct: HbProductEntity;
-    let secondProduct: HbProductEntity;
-
-    beforeEach(async () => {
-      const fixtures = async (_: Connection) => {
-        const model = await getModelFromEntity(HbProductEntity);
-        await model.insertMany([
-          { name: 'mayo', inStock: true },
-          { name: 'ketchup', inStock: false },
-        ]).then(([mayoProduct, ketchupProduct]) => {
-          firstProduct = mayoProduct as unknown as HbProductEntity;
-          secondProduct = ketchupProduct as unknown as HbProductEntity;
-        });
-      };
-
-      await initApp(
-        {
-          entity: HbProductEntity,
-          controllerOptions: {
-            apiTag: 'HbProduct',
-            path: 'hb-products',
-            isPublic: true,
-          },
-          routes: [
-            { type: 'GetMany', isPublic: true },
-            {
-              type: 'CreateOne',
-              broadcast: { enabled: true },
-            },
-            {
-              type: 'CreateMany',
-              broadcast: { enabled: (data: HbProductEntity) => data.inStock === true },
-            },
-            {
-              type: 'UpdateOne',
-              broadcast: { enabled: true, eventName: customEvent },
-            },
-            {
-              type: 'UpdateMany',
-              broadcast: { enabled: false },
-            },
-            {
-              type: 'ReplaceOne',
-              broadcast: { enabled: true },
-            },
-            {
-              type: 'DuplicateOne',
-              broadcast: { enabled: true },
-            },
-            {
-              type: 'DuplicateMany',
-              broadcast: { enabled: true },
-            },
-            {
-              type: 'DeleteOne',
-              broadcast: { enabled: true },
-            },
-            {
-              type: 'DeleteMany',
-              broadcast: { enabled: true },
-            },
-          ],
-        },
-        undefined,
-        fixtures,
-        async (app: INestApplication) => {
-          app.useWebSocketAdapter(new TestSocketAdapter(app));
-        },
-        true,
-      );
-    });
-
-    it('[CreateOne] should broadcast created product to all WS clients after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'post',
-        '/hb-products',
-        { name: 'mustard', inStock: true },
-        { broadcastEvent: 'create-one-hb-product' },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'create-one-hb-product',
-        data: expect.arrayContaining([
-          expect.objectContaining({ name: 'mustard', inStock: true }),
-        ]),
-      });
-      expect(broadcastData).toEqual(expect.arrayContaining([
-        expect.objectContaining({ name: 'mustard', inStock: true }),
-      ]));
-    });
-
-    it('[CreateMany] should broadcast only inStock products after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'post',
-        '/hb-products/many',
-        { list: [{ name: 'pickles', inStock: true }, { name: 'onions', inStock: false }] },
-        { broadcastEvent: 'create-many-hb-product' },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(broadcastData).toEqual(
-        expect.arrayContaining([expect.objectContaining({ name: 'pickles', inStock: true })]),
-      );
-      expect(broadcastData).not.toEqual(
-        expect.arrayContaining([expect.objectContaining({ name: 'onions' })]),
-      );
-    });
-
-    it('[UpdateOne] should broadcast updated product with custom event name after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'patch',
-        `/hb-products/${firstProduct.id}`,
-        { name: 'updated mayo', inStock: false },
-        { broadcastEvent: customEvent },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: customEvent,
-        data: expect.arrayContaining([
-          expect.objectContaining({ id: firstProduct.id, name: 'updated mayo', inStock: false }),
-        ]),
-      });
-      expect(broadcastData).toEqual(expect.arrayContaining([
-        expect.objectContaining({ name: 'updated mayo' }),
-      ]));
-    });
-
-    it('[ReplaceOne] should broadcast replaced product after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'put',
-        `/hb-products/${firstProduct.id}`,
-        { name: 'replaced mayo', inStock: false },
-        { broadcastEvent: 'replace-one-hb-product' },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(broadcastData).toEqual(expect.arrayContaining([
-        expect.objectContaining({ id: firstProduct.id, name: 'replaced mayo', inStock: false }),
-      ]));
-    });
-
-    it('[DuplicateOne] should broadcast duplicated product after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'post',
-        `/hb-products/duplicate/${firstProduct.id}`,
-        { name: 'duplicated mayo' },
-        { broadcastEvent: 'duplicate-one-hb-product' },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(broadcastData).toEqual(expect.arrayContaining([
-        expect.objectContaining({ name: 'duplicated mayo', inStock: firstProduct.inStock }),
-      ]));
-    });
-
-    it('[DuplicateMany] should broadcast duplicated products after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'post',
-        '/hb-products/duplicate',
-        { status: 'dup' },
-        {
-          broadcastEvent: 'duplicate-many-hb-product',
-          query: { ids: [firstProduct.id, secondProduct.id] },
-        },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(broadcastData.length).toBeGreaterThanOrEqual(2);
-    });
-
-    it('[DeleteOne] should broadcast deleted product id to all WS clients after HTTP call', async () => {
-      const { broadcastData } = await server.httpWithBroadcast(
-        'delete',
-        `/hb-products/${firstProduct.id}`,
-        {},
-        { broadcastEvent: 'delete-one-hb-product' },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(handleSocketBroadcast).toHaveBeenCalledWith({
-        event: 'delete-one-hb-product',
-        data: expect.arrayContaining([expect.objectContaining({ id: firstProduct.id })]),
-      });
-      expect(broadcastData).toEqual(expect.arrayContaining([{ id: firstProduct.id }]));
-    });
-
-    it('[DeleteMany] should broadcast deleted product ids to all WS clients after HTTP call', async () => {
-      const ids = [firstProduct.id, secondProduct.id];
-      const { broadcastData } = await server.httpWithBroadcast(
-        'delete',
-        '/hb-products',
-        {},
-        {
-          broadcastEvent: 'delete-many-hb-product',
-          query: { ids },
-        },
-      );
-
-      expect(handleSocketBroadcast).toHaveBeenCalledTimes(1);
-      expect(broadcastData).toEqual(expect.arrayContaining(ids.map(id => ({ id }))));
-    });
-
-    it('[UpdateMany] should NOT broadcast when enabled is false', async () => {
-      const { body: products } = await server.get('/hb-products');
-      const ids = products.map((p: HbProductEntity) => p.id);
-
-      const response = await server.patch(
-        `/hb-products/many?ids=${ids.join('&ids=')}`,
-        { status: 'updated' },
-      );
-
-      expect(response).toBeDefined();
-      expect(handleSocketBroadcast).not.toHaveBeenCalled();
-    });
-  });
-});
